requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
numpy==1.26.0